Validator fixed with main Controller class

+++ src/DBLayer/DBWarehouse.java
@@ -13,16 +13,16 @@ public class DBWarehouse implements IDBWarehouse{
      * @return ModelLayer.Warehouse
      * @throws SQLException: if something goes wrong
      */
-    public synchronized Warehouse create(float length, float width, float height) throws SQLException{
+    public synchronized Warehouse create(int length, int width, int height) throws SQLException{
         Warehouse warehouse = new Warehouse(length,width,height);
         try{
             java.sql.Connection conn = DBConnection.getInstance().getDBcon();
             PreparedStatement preparedStatement = conn.prepareStatement("INSERT INTO Warehouse " +
                     "(length, width, height) VALUES " +
                     "(?,?,?)");
-            preparedStatement.setFloat(1,length);
-            preparedStatement.setFloat(2,width);
-            preparedStatement.setFloat(3,height);
+            preparedStatement.setInt(1,length);
+            preparedStatement.setInt(2,width);
+            preparedStatement.setInt(3,height);
             preparedStatement.executeUpdate();
             String sql2 = "SELECT TOP 1 id FROM warehouse ORDER BY id DESC";
             ResultSet rs = conn.createStatement().executeQuery(sql2);
@@ -87,7 +87,7 @@ public class DBWarehouse implements IDBWarehouse{
      * @throws SQLException
      */
     private Warehouse buildObject(ResultSet rs) throws SQLException{
-        Warehouse warehouse = new Warehouse(rs.getFloat("length"), rs.getFloat("height"), rs.getFloat("width"));
+        Warehouse warehouse = new Warehouse(rs.getInt("length"), rs.getInt("height"), rs.getInt("width"));
         try {
             warehouse.setId(rs.getInt("id"));
         } catch(SQLException e) {
@@ -108,9 +108,9 @@ public class DBWarehouse implements IDBWarehouse{
         try {
             java.sql.Connection conn = DBConnection.getInstance().getDBcon();
             PreparedStatement preparedStatement = conn.prepareStatement("UPDATE Warehouse SET length=?, width=?, height=? WHERE id=?");
-            preparedStatement.setFloat(1,warehouse.getLength());
-            preparedStatement.setFloat(2,warehouse.getWidth());
-            preparedStatement.setFloat(3,warehouse.getHeight());
+            preparedStatement.setInt(1,warehouse.getLength());
+            preparedStatement.setInt(2,warehouse.getWidth());
+            preparedStatement.setInt(3,warehouse.getHeight());
             preparedStatement.setInt(4,warehouse.getId());
             preparedStatement.executeUpdate();
         } catch(SQLException e) {
